Extract import of realm-level and client-level composites in separate classes
