+++ src/main/java/de/adorsys/keycloak/config/service/rolecomposites/realm/ClientCompositeImport.java
@@ -1,0 +1,82 @@
+package com.github.borisskert.keycloak.config.service.rolecomposites.realm;
+
+import com.github.borisskert.keycloak.config.repository.ClientRepository;
+import com.github.borisskert.keycloak.config.repository.RoleRepository;
+import org.keycloak.representations.idm.RoleRepresentation;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Service;
+
+import java.util.*;
+import java.util.stream.Collectors;
+
+@Service("realmRoleClientCompositeImport")
+public class ClientCompositeImport {
+    private static final Logger logger = LoggerFactory.getLogger(ClientCompositeImport.class);
+
+    private final RoleRepository roleRepository;
+    private final ClientRepository clientRepository;
+
+    public ClientCompositeImport(RoleRepository roleRepository, ClientRepository clientRepository) {
+        this.roleRepository = roleRepository;
+        this.clientRepository = clientRepository;
+    }
+
+    public void update(String realm, String realmRole, Map<String, List<String>> clientComposites) {
+        for (Map.Entry<String, List<String>> clientCompositesByClients : clientComposites.entrySet()) {
+            String clientId = clientCompositesByClients.getKey();
+            List<String> clientCompositesByClient = clientCompositesByClients.getValue();
+
+            updateClientComposites(realm, realmRole, clientId, clientCompositesByClient);
+        }
+
+        removeRealmRoleClientComposites(realm, realmRole, clientComposites);
+    }
+
+    private void updateClientComposites(String realm, String realmRole, String clientId, List<String> Composites) {
+        List<String> existingClientCompositeNames = findRealmRoleClientCompositeNames(realm, realmRole, clientId);
+
+        if (Objects.equals(existingClientCompositeNames, Composites)) {
+            logger.debug("No need to update client-level role '{}'s composites client-roles for client '{}' in realm '{}'", realmRole, clientId, realm);
+        } else {
+            logger.debug("Update client-level role '{}'s composites client-roles for client '{}' in realm '{}'", realmRole, clientId, realm);
+
+            removeRealmRoleClientComposites(realm, realmRole, clientId, existingClientCompositeNames, Composites);
+            addRealmRoleClientComposites(realm, realmRole, clientId, existingClientCompositeNames, Composites);
+        }
+    }
+
+    private List<String> findRealmRoleClientCompositeNames(String realm, String realmRole, String clientId) {
+        Set<RoleRepresentation> existingClientComposites = roleRepository.findRealmRoleClientComposites(realm, realmRole, clientId);
+
+        return existingClientComposites.stream()
+                .map(RoleRepresentation::getName)
+                .collect(Collectors.toList());
+    }
+
+    private void removeRealmRoleClientComposites(String realm, String realmRole, String clientId, Collection<String> existingClientCompositeNames, Collection<String> clientCompositesByClient) {
+        Set<String> clientRoleCompositesToRemove = existingClientCompositeNames.stream()
+                .filter(name -> !clientCompositesByClient.contains(name))
+                .collect(Collectors.toSet());
+
+        roleRepository.removeRealmRoleClientComposites(realm, realmRole, clientId, clientRoleCompositesToRemove);
+    }
+
+    private void removeRealmRoleClientComposites(String realm, String realmRole, Map<String, List<String>> clientComposites) {
+        Set<String> existingCompositeClients = clientRepository.getClientIds(realm);
+
+        Set<String> compositeClientsToRemove = existingCompositeClients.stream()
+                .filter(name -> !clientComposites.containsKey(name))
+                .collect(Collectors.toSet());
+
+        roleRepository.removeRealmRoleClientComposites(realm, realmRole, compositeClientsToRemove);
+    }
+
+    private void addRealmRoleClientComposites(String realm, String realmRole, String clientId, Collection<String> existingClientCompositeNames, Collection<String> clientCompositesByClient) {
+        Set<String> clientRoleCompositesToAdd = clientCompositesByClient.stream()
+                .filter(name -> !existingClientCompositeNames.contains(name))
+                .collect(Collectors.toSet());
+
+        roleRepository.addRealmRoleClientComposites(realm, realmRole, clientId, clientRoleCompositesToAdd);
+    }
+}
