+++ src/main/java/de/adorsys/keycloak/config/service/rolecomposites/realm/RealmRoleCompositeImportService.java
@@ -1,0 +1,63 @@
+package com.github.borisskert.keycloak.config.service.rolecomposites.realm;
+
+import com.github.borisskert.keycloak.config.model.RealmImport;
+import org.keycloak.representations.idm.RoleRepresentation;
+import org.keycloak.representations.idm.RolesRepresentation;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
+
+/**
+ * Implements the update mechanism for role composites of realm-level roles
+ */
+@Service
+public class RealmRoleCompositeImportService {
+
+    private final RealmCompositeImport realmCompositeImport;
+    private final ClientCompositeImport clientCompositeImport;
+
+    @Autowired
+    public RealmRoleCompositeImportService(
+            RealmCompositeImport realmCompositeImport,
+            ClientCompositeImport clientCompositeImport
+    ) {
+        this.clientCompositeImport = clientCompositeImport;
+        this.realmCompositeImport = realmCompositeImport;
+    }
+
+    /**
+     * Updates the role composites for all realm-level roles
+     *
+     * @param realmImport the realm-import containing all realm-level roles containing role-composites to be imported
+     */
+    public void update(RealmImport realmImport) {
+        String realm = realmImport.getRealm();
+        RolesRepresentation roles = realmImport.getRoles();
+        List<RoleRepresentation> realmRoles = roles.getRealm();
+
+        for (RoleRepresentation realmRole : realmRoles) {
+            updateRealmRoleRealmCompositesIfNecessary(realm, realmRole);
+            updateRealmRoleClientCompositesIfNecessary(realm, realmRole);
+        }
+    }
+
+    private void updateRealmRoleRealmCompositesIfNecessary(String realm, RoleRepresentation realmRole) {
+        Optional.ofNullable(realmRole.getComposites())
+                .flatMap(composites -> Optional.ofNullable(composites.getRealm()))
+                .ifPresent(realmComposites -> {
+                    realmCompositeImport.update(realm, realmRole, realmComposites);
+                });
+    }
+
+    private void updateRealmRoleClientCompositesIfNecessary(String realm, RoleRepresentation realmRole) {
+        Optional.ofNullable(realmRole.getComposites())
+                .flatMap(composites -> Optional.ofNullable(composites.getClient()))
+                .ifPresent(clientComposites -> clientCompositeImport.update(
+                        realm,
+                        realmRole.getName(),
+                        clientComposites
+                ));
+    }
+}
